[FEAT] Update UI
